feat(edititemform, dropdown): form in singleItem has been migrated to editSingleItem
drop down was also created. SUbmit has a but on edit single item, not ready yet

BREAKING CHANGE: see bug note above
